chore: fix collisions + serialization

+++ Phoenix/Core/ECS/include/EntityComponent.h
@@ -3,7 +3,7 @@
  {
   using ComponentId = long long int;
   using ComponentSystemId = size_t;\
-  using EntityId = size_t;
+  using EntityId = long long int;
   // EntityIdExtended can be used for systems that need to store the entity id with unset values like -1
   using EntityIdExtended = long long int;
  }
